slow down full sim premixing step
