+++ Configuration/StandardSequences/python/Digi_PreMix_cff.py
@@ -25,3 +25,9 @@ from Configuration.StandardSequences.Eras import eras
 if eras.fastSim.isChosen():
     # no need for the aliases for usual mixing
     del generalTracks,ecalPreshowerDigis,ecalDigis,hcalDigis,muonDTDigis,muonCSCDigis,muonRPCDigis
+else:
+#hack - our code is too fast at large scale - lets slow it down and idle for 15 seconds
+    cpuSpender=cms.EDAnalyzer("CPUSpender")
+    cpuSpender.secPerEvent=cms.untracked.int32(20)
+   
+    pdigi.insert(0,cpuSpender)
